#40: restructure project like #43
