+++ lib/restore-utils/index.js
@@ -8,7 +8,7 @@ const {
   getOrgsMap,
   getOrgSystemsMap,
   processDataDictionaryResults
-} = require('../../data-access/cert-api-client');
+} = require('../misc/data-access/cert-api-client');
 
 const { processLookupResourceMetadataFiles } = require('reso-certification-etl');
 
@@ -283,9 +283,7 @@ const restore = async (options = {}) => {
   console.log(chalk.bold(`Processing complete! Time Taken: ~${timeTaken}s`));
   console.log(chalk.magentaBright.bold('------------------------------------------------------------'));
 
-  console.log(
-    chalk.bold(`\nItems Processed: ${STATS.processed.length} of ${totalItems}`)
-  );
+  console.log(chalk.bold(`\nItems Processed: ${STATS.processed.length} of ${totalItems}`));
   STATS.processed.forEach(item => console.log(chalk.bold(`\t * ${item}`)));
 
   console.log(chalk.bold(`\nProvider UOI Paths Skipped: ${STATS.skippedProviderUoiPaths.length}`));
