+++ lib/schema/generate.js
