+++ lib/schema/index.js
