+++ lib/schema/validate.js
